Fix JS arrays

+++ src/gfx-ui/SpaceSheet.cpp
@@ -47,7 +47,7 @@ namespace mud
 		, m_model("SpaceSheet")
 		, m_item(m_node, m_model, ItemFlag::Ui, &m_material)
 	{
-		load_texture_rgba(m_texture, uint16_t(width), uint16_t(height), span<uint8_t>{});
+		load_texture_rgba(m_texture, uint16_t(width), uint16_t(height), span<uint32_t>{});
 		m_material.m_program = viewport.m_scene->m_gfx.programs().file("debug");
 		m_material.m_solid.m_colour = &m_texture;
